Fix InvalidArgument exception thrown in FrameProvider sample during service shutdown
This fix works-around the behavior in SensorDataService when the service
is  shutting down, in which the FrameProvider object is directly Closed
by the service while ProviderManager still references the Provider
object.

The ProviderManager no longer releases and cleans-up the Provider when
the Destructor is invoked; it assumes this has already been performed by
SensorDataService. Also, this change adds additional thread locks to
protect against a potential race-condition if the camera is physically
disconnected at the same time the service is shutting down.